less type casts

+++ tsdb/chunks/head_chunks.go
@@ -162,7 +162,7 @@ func (f *chunkPos) shouldCutNewFile(chunkSize uint64) bool {
 // Head chunk format: https://github.com/prometheus/prometheus/blob/main/tsdb/docs/format/head_chunks.md#chunk
 func (f *chunkPos) bytesToWriteForChunk(chkLen uint64) uint64 {
 	// headers
-	bytes := uint64(SeriesRefSize) + 2*uint64(MintMaxtSize) + uint64(ChunkEncodingSize)
+	bytes := uint64(SeriesRefSize) + 2*MintMaxtSize + ChunkEncodingSize
 
 	// size of chunk length encoded as uvarint
 	bytes += f.sizeAsUVarInt(chkLen)
